+++ src/settings/NoticePolicy/components/DetailSections/components/NoticeCard/NoticeCard.test.js
@@ -12,6 +12,7 @@ import {
   loanNoticesTriggeringEvents,
   noticesSendEvents,
 } from '../../../../../../constants';
+// import { REAL_TIME_NOTICES_TRANSLATION_KEYS } from '../../../../utils/get-real-time-labels';
 
 const POLICY_NOTICES_TRANSLATION_KEY = 'ui-circulation.settings.noticePolicy.notices';
 const testIds = {
@@ -34,6 +35,11 @@ const testIds = {
   noticeCardSendEveryIntervalId: 'noticeCardSendEveryIntervalId',
   header: 'header',
 };
+
+jest.mock('../../../../utils/get-real-time-labels', () => jest.fn(() => ({
+  longTerm: 'ui-circulation.settings.noticePolicy.notices.send.longTerm',
+  shortTerm: 'ui-circulation.settings.noticePolicy.notices.send.shortTerm',
+})));
 
 const testKeyValues = (testId, label, value) => {
   if (!value) {
@@ -72,6 +78,7 @@ describe('View NoticeCard in DetailSection', () => {
         isFrequencyAvailable: jest.fn(() => true),
         isLoanDueDateTimeSelected: jest.fn(() => true),
         isSendOptionsAvailable: jest.fn(() => true),
+        isLostItemFeesSelected: jest.fn(() => true),
         sendWhen: 'Check out',
       },
       templateId: 'testTemplateId',
@@ -183,6 +190,7 @@ describe('View NoticeCard in DetailSection', () => {
           isBeforeOrAfter: jest.fn(() => false),
           isFrequencyAvailable: jest.fn(() => false),
           isLoanDueDateTimeSelected: jest.fn(() => false),
+          isLostItemFeesSelected: jest.fn(() => false),
         },
       },
     };
@@ -199,7 +207,7 @@ describe('View NoticeCard in DetailSection', () => {
       testNotShownValues(testIds.noticeCardFrequencyLabel);
     });
 
-    describe('when "isLoanDueDateTimeSelected" is "false"', () => {
+    describe('when "isLoanDueDateTimeSelected" and "isLostItemFeesSelected" are "false"', () => {
       it('should not render real time message', () => {
         expect(screen.queryByText(`${POLICY_NOTICES_TRANSLATION_KEY}.send.shortTerm`))
           .not.toBeInTheDocument();